Crate is_ci

Statics

INIT 🔒

IS_CI 🔒

Functions

cached

Returns true if the current environment is found to probably be a CI environment or service, and caches the result for future calls. If you expect the environment to change, use uncached.

check 🔒

is_ciDeprecated

Returns true if the current environment is found to probably be a CI environment or service. That’s it, that’s all it does.

uncached

Returns true if the current environment is found to probably be a CI environment or service. If you expect to call this multiple times without the environment changing, use cached.